requirements.txt
flask==3.1.1
flask-cors==5.0.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.41
python-dotenv==1.1.0
flask-sqlalchemy==3.1.1
marshmallow==3.26.1
flask-marshmallow==1.2.1
werkzeug==3.1.3
